fix login/logout, add protected routes
hopefully this fixes login/logout once and for all. i hoisted the user
state up to the app level, where it's now apparent it really belongs.
there's one remaining hole: we don't properly handle session expiry.
when the session expires, the app won't notice. the only thing that'll
happen is that the grpcs will start failing.