simplify itemlist interface
1. replace mutable and showclaim with mode
2. don't instantiate itemlistelement using an API - do it directly

the first is because itemlist's needs were becoming more complicated
than the two mode bits could handle. the second because it seemed like a
good idea at the time, when there was going to be more edit- and
view-specific stuff in their respective classes, but that turns out not
to be the case.