don't put claimuser in the listitem
this seemed like a good idea at the time, but added all sorts of
complication for ListItem construction. as it happens, nobody needs the
claim user name synchronously (the claim chip is the only thing that needs it
at all), and the async cases can get it asynchronously.

putting the claim user in the item also makes it difficult to have the
native item object be the object returned by the list model (without
merging the list and user models).

the only worrisome case is loading a large list, where there could
conceivably be a bunch of requests for users all at the same time. the
pre-loading of users at list load time should eliminate that though